Review comments, renames + comments.
