+++ lib/iris/fileformats/netcdf.py
@@ -44,7 +44,7 @@ import iris.cube
 import iris.exceptions
 import iris.fileformats.cf
 import iris.fileformats._pyke_rules
-from iris.fileformats._ugrid_cf_reader import UGridCFReader
+from iris.fileformats.ugrid_cf_reader import UGridCFReader
 import iris.io
 import iris.util
 from iris._lazy_data import as_lazy_data
@@ -778,18 +778,20 @@ def load_cubes(filenames, callback=None, *args, **kwargs):
         filenames = [filenames]
 
     for filename in filenames:
-        # Ingest the netCDF file.
-        #         ncreader = iris.fileformats.cf.CFReader(filename)
-        ncreader = UGridCFReader(filename, *args, **kwargs)
+        # Ingest the netCDF file, creating a reader which also checks for UGRID
+        # content.
+        reader = UGridCFReader(filename, *args, **kwargs)
 
         # Process each CF data variable.
         data_variables = list(
-            ncreader.cfreader.cf_group.data_variables.values()
-        ) + list(ncreader.cfreader.cf_group.promoted.values())
+            reader.cfreader.cf_group.data_variables.values()
+        ) + list(reader.cfreader.cf_group.promoted.values())
         for cf_var in data_variables:
-            cube = _load_cube(engine, ncreader.cfreader, cf_var, filename)
+            cube = _load_cube(engine, reader.cfreader, cf_var, filename)
 
-            ncreader.complete_ugrid_cube(cube)
+            # Post-process each cube to attach information describing the
+            # unstructured mesh dimension, if any.
+            reader.complete_ugrid_cube(cube)
 
             # Process any associated formula terms and attach
             # the corresponding AuxCoordFactory.
